done pdf

+++ lib/app/drawer.dart
@@ -1,4 +1,6 @@
 import 'package:bsthrm/app/common/loginpage.dart';
+import 'package:bsthrm/app/page/profile/id_card.dart';
+import 'package:bsthrm/app/page/salary/salary_slip.dart';
 import 'package:bsthrm/app/page/verify_pages/aadhar_verify.dart';
 import 'package:bsthrm/app/page/verify_pages/bank_verify.dart';
 import 'package:bsthrm/app/page/verify_pages/kyc_settings.dart';
@@ -8,6 +10,7 @@ import 'package:bsthrm/global/fontstyle.dart';
 import 'package:bsthrm/viewmodel/cubit/app_state_cubit.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:nb_utils/nb_utils.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 class AppDrawer extends StatefulWidget {
@@ -32,33 +35,31 @@ class _AppDrawerState extends State<AppDrawer> {
             child: SafeArea(
               child: Padding(
                 padding: const EdgeInsets.all(8.0),
-                child: Flexible(
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      SizedBox(
-                        height: 60,
-                        child: Image.network(
-                          appState.userDetails?.employeePhoto! ?? '',
-                        ),
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    SizedBox(
+                      height: 60,
+                      child: Image.network(
+                        appState.userDetails?.employeePhoto! ?? '',
                       ),
-                      Text(
-                        appState.userDetails?.employeeName ?? '',
-                        style: hsSemiBold.copyWith(
-                            fontSize: 20, color: Colors.black),
-                      ),
-                      Text(
-                        "(${appState.userDetails?.employeeDesignation})" ?? '',
-                        style: hsSemiBold.copyWith(
-                            fontSize: 18, color: Colors.black),
-                      ),
-                      Text(
-                        appState.userDetails?.empid ?? '',
-                        style: hsSemiBold.copyWith(
-                            fontSize: 16, color: Colors.black),
-                      ),
-                    ],
-                  ),
+                    ),
+                    Text(
+                      appState.userDetails?.employeeName ?? '',
+                      style: hsSemiBold.copyWith(
+                          fontSize: 20, color: Colors.black),
+                    ),
+                    Text(
+                      "(${appState.userDetails?.employeeDesignation})" ?? '',
+                      style: hsSemiBold.copyWith(
+                          fontSize: 18, color: Colors.black),
+                    ),
+                    Text(
+                      appState.userDetails?.empid ?? '',
+                      style: hsSemiBold.copyWith(
+                          fontSize: 16, color: Colors.black),
+                    ),
+                  ],
                 ),
               ),
             ),
@@ -134,22 +135,56 @@ class _AppDrawerState extends State<AppDrawer> {
                   );
                 },
               ),
+              ListTile(
+                title: const Text('Salary Slip'),
+                onTap: () {
+                  Navigator.pop(context);
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                      builder: (context) => const SalarySlip(),
+                    ),
+                  );
+                },
+              ),
             ],
           ),
           ListTile(
             leading: const Icon(Icons.logout),
             title: const Text('Logout'),
             onTap: () async {
-              var navigator = Navigator.of(context);
-              SharedPreferences preferences =
-                  await SharedPreferences.getInstance();
-              await preferences.clear();
-              debugPrint('All SharedPreferences values removed');
-              navigator.push(MaterialPageRoute(
-                builder: (context) {
-                  return const LoginPage();
-                },
-              ));
+              //  show alert dialod for confirm
+              showDialog(
+                  context: context,
+                  builder: (BuildContext context) {
+                    return AlertDialog(
+                      title: const Text('Logout'),
+                      content: const Text('Are you sure you want to logout?'),
+                      actions: [
+                        TextButton(
+                          onPressed: () async {
+                            var navigator = Navigator.of(context);
+                            SharedPreferences preferences =
+                                await SharedPreferences.getInstance();
+                            await preferences.clear();
+                            debugPrint('All SharedPreferences values removed');
+                            navigator.push(MaterialPageRoute(
+                              builder: (context) {
+                                return const LoginPage();
+                              },
+                            ));
+                          },
+                          child: const Text('Yes'),
+                        ),
+                        TextButton(
+                          onPressed: () {
+                            Navigator.pop(context);
+                          },
+                          child: const Text('No'),
+                        ),
+                      ],
+                    );
+                  });
             },
           ),
         ],
